pyproject.toml
[project]
name = "finpipe-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "massive>=2.3.2",
    "python-dotenv>=1.2.2",
    "rich>=13.0.0",
    "uvicorn>=0.34.0",
]
